feat: adicionando autenticação no back end e no front end e integrando a api

+++ server/src/route/memories.ts
@@ -1,11 +1,22 @@
 import { FastifyInstance } from "fastify";
 import { prisma } from "../lib/prisma";
-import { z } from "zod"
+import { z } from "zod";
  
 export async function memoriesRoute(app:FastifyInstance) {
 
-    app.get("/memories", async() => {
+    app.addHook("preHandler", async (request) => {
+        await request.jwtVerify()
+    })
+
+    app.get("/memories", async(request) => {
+
+       //const {} 
+
         const memories = await prisma.memory.findMany({
+            where:{
+                userId: request.user.sub,
+            },
+
             orderBy : { 
                 createdAt: "asc",
             },
@@ -20,7 +31,7 @@ export async function memoriesRoute(app:FastifyInstance) {
         })
     });
 
-    app.get("/memories/:id", async(request) => {
+    app.get("/memories/:id", async(request, reply) => {
         const paramsSchema = z.object({
             id: z.string().uuid(),
         })
@@ -32,6 +43,12 @@ export async function memoriesRoute(app:FastifyInstance) {
                 id,
             }, 
         })
+
+        if(!memory.isPublic && memory.userId !== request.user.sub){
+            return reply.status(401).send()
+        }
+
+
         return memory
          
     });
@@ -46,12 +63,12 @@ export async function memoriesRoute(app:FastifyInstance) {
 
         const{ content, coverUrl ,isPublic } = bodySchema.parse(request.body)
 
-        const memory = await prisma.memory.create({
+        const memory = await prisma.memory.create({ 
             data: {
                 content,
                 coverUrl,
                 isPublic,
-                userId: "e7ff4f44-f325-4ec1-9244-7701c5fabd90", 
+                userId: request.user.sub, 
             },
         })
         
@@ -59,7 +76,7 @@ export async function memoriesRoute(app:FastifyInstance) {
        
     }); 
 
-    app.put("/memories/:id", async(request) => {
+    app.put("/memories/:id", async(request, reply) => {
         const paramsSchema = z.object({
             id: z.string().uuid(),
         })
@@ -74,7 +91,17 @@ export async function memoriesRoute(app:FastifyInstance) {
 
         const{ content, coverUrl ,isPublic } = bodySchema.parse(request.body)
 
-      const memory =  await prisma.memory.update({
+        let memory = await prisma.memory.findUniqueOrThrow({
+            where:{
+                id,
+            }
+        })
+
+        if(memory.userId !== request.user.sub){
+            return reply.status(401).send()
+        }
+
+       memory =  await prisma.memory.update({
             where: {
                 id, 
             },
@@ -90,12 +117,23 @@ export async function memoriesRoute(app:FastifyInstance) {
         
     });
 
-    app.delete("/memories/:id", async(request) => {
+    app.delete("/memories/:id", async(request, reply) => {
         const paramsSchema = z.object({
             id: z.string().uuid(),
         })
 
         const { id } = paramsSchema.parse(request.params)
+
+        const memory = await prisma.memory.findUniqueOrThrow({
+            where:{
+                id,
+            }
+        })
+
+        if(memory.userId !== request.user.sub){
+            return reply.status(401).send()
+        }
+
 
          await prisma.memory.delete({
             where:{
